Moved around functions/classes to correct files
